new bd schema

+++ database/queries.py
@@ -1,8 +1,5 @@
 from database.connection import DataBase
 from Crypto.Hash import SHA256
-import os
-import random
-import string
 from utils.jwt import JWTUtils
 
 
@@ -19,32 +16,32 @@ class SQLQueriesTasks():
 
         self.database.create_table_tasks(self.table_name)
 
-    def create_new_row(self, name, pomodoro_count, category_id):
-        query = "INSERT INTO {table_name} (name, pomodoro_count, category_id) VALUES (%s, %s, %s) RETURNING id"
-        self.cursor.execute(query.format(table_name=self.table_name), (name, pomodoro_count, category_id))
+    def create_new_row(self, name, pomodoro_count, category_id, user_id):
+        query = "INSERT INTO {table_name} (name, pomodoro_count, category_id, user_id) VALUES (%s, %s, %s, %s) RETURNING id"
+        self.cursor.execute(query.format(table_name=self.table_name), (name, pomodoro_count, category_id, user_id))
         self.connection.commit()
         new_row_id = self.cursor.fetchone()[0]
         return new_row_id
 
-    def update_task_name(self, task_id, new_name):
-        query = "UPDATE {table_name} SET name = %s WHERE id = %s"
-        self.cursor.execute(query.format(table_name=self.table_name), (new_name, task_id))
+    def update_task_name(self, task_id, new_name, user_id):
+        query = "UPDATE {table_name} SET name = %s WHERE (id = %s and user_id = %s)"
+        self.cursor.execute(query.format(table_name=self.table_name), (new_name, task_id, user_id))
         self.connection.commit()
-        query = "SELECT * FROM {table_name} WHERE id = %s"
-        self.cursor.execute(query.format(table_name=self.table_name), (task_id,))
+        query = "SELECT * FROM {table_name} WHERE (id = %s and user_id = %s)"
+        self.cursor.execute(query.format(table_name=self.table_name), (task_id, user_id))
         rows = self.cursor.fetchall()
         return rows
 
-    def delete_row_by_id(self, task_id):
-        query = "DELETE FROM {table_name} WHERE id = %s"
-        self.cursor.execute(query.format(table_name=self.table_name), (task_id,))
+    def delete_row_by_id(self, task_id, user_id):
+        query = "DELETE FROM {table_name} WHERE (id = %s and user_id = %s)"
+        self.cursor.execute(query.format(table_name=self.table_name), (task_id, user_id))
         deleted_count = self.cursor.rowcount
         self.connection.commit()
         return deleted_count > 0
 
-    def select_all_rows(self):
-        query = "SELECT * FROM {table_name}"
-        self.cursor.execute(query.format(table_name=self.table_name))
+    def select_all_rows(self, user_id):
+        query = "SELECT * FROM {table_name} WHERE user_id = %s"
+        self.cursor.execute(query.format(table_name=self.table_name), (user_id,))
         rows = self.cursor.fetchall()
         return rows
 
@@ -123,19 +120,6 @@ class SQLQueriesUsers():
         # Возвращаем хэш в виде шестнадцатеричной строки
         return sha256_hash.hexdigest()
     
-    @staticmethod
-    def generate_access_token():
-        # Определяем символы, которые могут быть использованы в токене
-        characters = string.ascii_letters + string.digits + string.punctuation
-        
-        # Генерируем случайное начальное число (seed) на основе системного времени
-        random.seed(os.urandom(32))
-        
-        # Генерируем 10-символьный токен
-        token = ''.join(random.choice(characters) for _ in range(10))
-        
-        return token
-
     def close(self):
         self.cursor.close()
         self.connection.close()
